+++ src/lib.rs
@@ -34,4 +34,4 @@ mod vec;
 pub use arena::ArenaStats;
 pub use arena::{Arena, Checkpoint};
 pub use util::{Transaction, TxnDiff, TxnStatus};
-pub use vec::{ArenaVec, TryReserveError};
+pub use vec::{ArenaVec, ArenaVecIntoIter, TryReserveError};